+++ CHANGELOG.md
@@ -19,3 +19,14 @@ Second alpha build. Extensive refactoring, and added the ability to compare file
     - Replaced deprecated synchronous filesystem functions with promises (and obviously awaited them, still need them to be synchronous!)
     - Fixed a few TypeScript warnings related to the registered mod list.
 - Started working on handling mod ancestry. Nothing actually functional yet, just dangling bits of code here and there for now.
+
+## 0.3.0
+
+- Cloning/comparing another mod's files now falls back to its ancestors (if any) before defaulting to the base game.
+    - This process only checks the registered mod list. If an ancestor's name has not been registered with the OpenSR Toolset yet, it will not be detected and the operation will fall through to the base game.
+- Added a third target type for clone/compare commands: Ancestor. Ancestor mode tries to clone or compare files from the current mod's ancestors (if any) or the base game.
+- Mod registration now reads the mod name from `modinfo.txt` instead of requiring the user to supply a name themselves.
+    - Added a `Reregister Mod` command to update the name of a previously registered mod.
+- More refactoring: The extension was getting a little bloated and hard to navigate, so I broke it up into separate files.
+- Added logging under `Output -> OpenSR Toolset`. This is subject to VSCode's log level, configurable via `Developer: Set Log Level...`.
+- Added a custom editor title for compare operations, denoting the source and destination mods for the comparison. (The base game is labeled `Base`).